// dashboard/src/models/interfaces/Subcategory.ts
/**
 * Interface for Subcategory model
 *
 * @interface Subcategory
 * @property {string} id - ID of the subcategory.
 * @property {string} name - Name of the subcategory.
 * @property {number} categoryId - ID of the category of the subcategory.
 */
export interface Subcategory {
  id: string;
  name: string;
  categoryId: string;
}

/**
 * Interface for Subcategory model.
 *
 * @interface NewSubcategory
 * @property {string} name - Name of the subcategory.
 * @property {number} categoryId - ID of the category of the subcategory.
 */
export interface NewSubcategory {
  name: string;
  categoryId: string;
}

/**
 * Interface for Subcategory model.
 *
 * @interface UpdateSubcategory
 * @property {string} name - Name of the subcategory.
 * @property {number} categoryId - ID of the category of the subcategory.
 */
export interface UpdateSubcategory extends NewSubcategory {

}